refactor: move base.ts to src/oclif-commands/
Move BaseCommand from src/commands/ to src/oclif-commands/
to consolidate all oclif code in one directory. Updates 16
import paths across command files.
