+++ src/oclif-commands/journal/meeting.ts
@@ -4,7 +4,7 @@ import process from 'node:process'
 import { Args } from '@oclif/core'
 import consola from 'consola'
 import { colors } from 'consola/utils'
-import { BaseCommand } from '../../commands/base.js'
+import { BaseCommand } from '../base.js'
 import { JOURNAL_TYPES } from '../../types/journal.js'
 import {
   createMeetingContent,
