+++ src/oclif-commands/observe/report.ts
@@ -3,7 +3,7 @@ import * as fs from 'node:fs'
 import * as os from 'node:os'
 import * as path from 'node:path'
 import { x } from 'tinyexec'
-import { BaseCommand } from '../../commands/base.js'
+import { BaseCommand } from '../base.js'
 
 /**
  * Generate token/cost usage report via ccusage
