+++ src/oclif-commands/observe/session.ts
@@ -3,7 +3,7 @@ import * as fs from 'node:fs'
 import * as os from 'node:os'
 import * as path from 'node:path'
 import pc from 'picocolors'
-import { BaseCommand } from '../../commands/base.js'
+import { BaseCommand } from '../base.js'
 
 interface JournalEntry {
   type: string
