+++ src/oclif-commands/ralph/task/add.ts
@@ -1,6 +1,6 @@
 import { Args, Flags } from '@oclif/core'
 import pc from 'picocolors'
-import { BaseCommand } from '../../../commands/base.js'
+import { BaseCommand } from '../../base.js'
 import {
   DEFAULT_STATE_FILE,
   DEFAULT_MAX_ITERATIONS,
